refac: rearragne calculation folder
