+++ source/databricks/calculation_engine/package/calculation/input/__init__.py
@@ -11,3 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+"""
+Package `input`
+
+Responsible for reading calculation input data from Delta tables.
+Schemas are checked when reading.
+Only minor transformations to data frames are done merely compensating
+for minor inappropriatenesses in the input table formats.
+"""
+
+from .table_reader import TableReader
