+++ source/databricks/calculation_engine/package/calculation/input/schemas/grid_loss_metering_points_schema.py
