+++ source/databricks/calculation_engine/package/calculation/output/basis_data_results.py
@@ -15,7 +15,7 @@ from dependency_injector.wiring import Provide, inject
 
 from package.calculation.CalculationResults import BasisDataContainer
 from package.calculation.calculator_args import CalculatorArgs
-from package.calculation_output.basis_data_writer import BasisDataWriter
+from package.calculation.output.basis_data_writer import BasisDataWriter
 from package.codelists import AggregationLevel
 from package.constants import PartitionKeyName
 from package.container import Container
