+++ source/databricks/calculation_engine/package/calculation/output/schemas/energy_results_schema.py
