+++ source/databricks/calculation_engine/package/calculation/preparation/transformations/grid_loss_responsible.py
@@ -22,7 +22,7 @@ from pyspark.sql.functions import col
 from package.calculation.preparation.grid_loss_responsible import (
     GridLossResponsible,
 )
-from package.calculation_input import TableReader
+from package.calculation.input import TableReader
 from package.codelists import MeteringPointType
 from package.constants import Colname
