+++ source/databricks/calculation_engine/package/calculation/preparation/transformations/time_series_points.py
@@ -17,10 +17,8 @@ from datetime import datetime
 from pyspark.sql import DataFrame
 from pyspark.sql.functions import col
 
-from package.calculation_input import TableReader
+from package.calculation.input import TableReader
 from package.constants import Colname
-from package.common import assert_schema
-from package.calculation_input.schemas import time_series_point_schema
 
 
 def get_time_series_points(
