+++ source/databricks/calculation_engine/tests/calculation/input/table_reader/input_metering_point_periods_factory.py
@@ -17,7 +17,7 @@ from datetime import datetime
 
 from pyspark.sql import DataFrame, Row, SparkSession
 
-from package.calculation_input.schemas import (
+from package.calculation.input import (
     metering_point_period_schema,
 )
 from package.codelists import (
