+++ source/databricks/calculation_engine/tests/calculation/input/table_reader/test_read_charge_master_data_periods.py
@@ -18,31 +18,38 @@ import pytest
 from pyspark.sql import SparkSession
 import pyspark.sql.functions as f
 
-from package.calculation_input.table_reader import TableReader
-from package.calculation_input.schemas import (
-    grid_loss_metering_points_schema,
-)
+
+from package.calculation.input.table_reader import TableReader
+from package.calculation.input import charge_master_data_periods_schema
 from package.constants import Colname
 from tests.helpers.delta_table_utils import write_dataframe_to_table
 from tests.helpers.data_frame_utils import assert_dataframes_equal
 
-DEFAULT_OBSERVATION_TIME = datetime(2022, 6, 8, 22, 0, 0)
 DEFAULT_FROM_DATE = datetime(2022, 6, 8, 22, 0, 0)
-DEFAULT_TO_DATE = datetime(2022, 6, 9, 22, 0, 0)
+DEFAULT_TO_DATE = datetime(2022, 6, 8, 22, 0, 0)
 
 
-def _create_grid_loss_metering_point_row() -> dict:
+def _create_charge_master_period_row() -> dict:
     return {
-        Colname.metering_point_id: "570715000000682292",
+        Colname.charge_code: "foo",
+        Colname.charge_type: "foo",
+        Colname.charge_owner: "foo",
+        Colname.resolution: "foo",
+        Colname.charge_tax: False,
+        Colname.from_date: DEFAULT_FROM_DATE,
+        Colname.to_date: DEFAULT_TO_DATE,
     }
 
 
 class TestWhenSchemaMismatch:
-    def test_raises_assertion_error(self, spark: SparkSession) -> None:
+    def test__raises_assertion_error(
+        self,
+        spark: SparkSession,
+    ) -> None:
         # Arrange
-        row = _create_grid_loss_metering_point_row()
+        row = _create_charge_master_period_row()
         reader = TableReader(mock.Mock(), "dummy_calculation_input_path")
-        df = spark.createDataFrame(data=[row], schema=grid_loss_metering_points_schema)
+        df = spark.createDataFrame(data=[row], schema=charge_master_data_periods_schema)
         df = df.withColumn("test", f.lit("test"))
 
         # Act & Assert
@@ -50,7 +57,7 @@ class TestWhenSchemaMismatch:
             reader._spark.read.format("delta"), "load", return_value=df
         ):
             with pytest.raises(AssertionError) as exc_info:
-                reader.read_grid_loss_metering_points()
+                reader.read_charge_master_data_periods()
 
             assert "Schema mismatch" in str(exc_info.value)
 
@@ -60,25 +67,26 @@ class TestWhenValidInput:
         self,
         spark: SparkSession,
         tmp_path: pathlib.Path,
+        calculation_input_folder: str,
     ) -> None:
         # Arrange
-        calculation_input_path = f"{str(tmp_path)}/calculation_input"
-        table_location = f"{calculation_input_path}/grid_loss_metering_points"
-        row = _create_grid_loss_metering_point_row()
-        df = spark.createDataFrame(data=[row], schema=grid_loss_metering_points_schema)
+        calculation_input_path = f"{str(tmp_path)}/{calculation_input_folder}"
+        table_location = f"{calculation_input_path}/charge_masterdata_periods"
+        row = _create_charge_master_period_row()
+        df = spark.createDataFrame(data=[row], schema=charge_master_data_periods_schema)
         write_dataframe_to_table(
             spark,
             df,
             "test_database",
-            "grid_loss_metering_points",
+            "charge_master_data_periods",
             table_location,
-            grid_loss_metering_points_schema,
+            charge_master_data_periods_schema,
         )
         expected = df
         reader = TableReader(spark, calculation_input_path)
 
         # Act
-        actual = reader.read_grid_loss_metering_points()
+        actual = reader.read_charge_master_data_periods()
 
         # Assert
         assert_dataframes_equal(actual, expected)
