+++ source/databricks/calculation_engine/tests/calculation/input/table_reader/test_read_charge_price_points.py
@@ -13,31 +13,31 @@
 # limitations under the License.
 import pathlib
 from datetime import datetime
+from decimal import Decimal
 from unittest import mock
 import pytest
-from pyspark.sql import SparkSession, DataFrame
+from pyspark.sql import SparkSession
 import pyspark.sql.functions as f
 
 
-from package.calculation_input.table_reader import TableReader
-from package.calculation_input.schemas import charge_master_data_periods_schema
+from package.calculation.input.table_reader import TableReader
+from package.calculation.input import charge_price_points_schema
 from package.constants import Colname
 from tests.helpers.delta_table_utils import write_dataframe_to_table
 from tests.helpers.data_frame_utils import assert_dataframes_equal
 
+DEFAULT_OBSERVATION_TIME = datetime(2022, 6, 8, 22, 0, 0)
 DEFAULT_FROM_DATE = datetime(2022, 6, 8, 22, 0, 0)
 DEFAULT_TO_DATE = datetime(2022, 6, 8, 22, 0, 0)
 
 
-def _create_charge_master_period_row() -> dict:
+def _create_change_price_point_row() -> dict:
     return {
         Colname.charge_code: "foo",
         Colname.charge_type: "foo",
         Colname.charge_owner: "foo",
-        Colname.resolution: "foo",
-        Colname.charge_tax: False,
-        Colname.from_date: DEFAULT_FROM_DATE,
-        Colname.to_date: DEFAULT_TO_DATE,
+        Colname.charge_price: Decimal("1.123456"),
+        Colname.charge_time: DEFAULT_OBSERVATION_TIME,
     }
 
 
@@ -47,9 +47,9 @@ class TestWhenSchemaMismatch:
         spark: SparkSession,
     ) -> None:
         # Arrange
-        row = _create_charge_master_period_row()
+        row = _create_change_price_point_row()
         reader = TableReader(mock.Mock(), "dummy_calculation_input_path")
-        df = spark.createDataFrame(data=[row], schema=charge_master_data_periods_schema)
+        df = spark.createDataFrame(data=[row], schema=charge_price_points_schema)
         df = df.withColumn("test", f.lit("test"))
 
         # Act & Assert
@@ -57,7 +57,7 @@ class TestWhenSchemaMismatch:
             reader._spark.read.format("delta"), "load", return_value=df
         ):
             with pytest.raises(AssertionError) as exc_info:
-                reader.read_charge_master_data_periods()
+                reader.read_charge_price_points()
 
             assert "Schema mismatch" in str(exc_info.value)
 
@@ -71,22 +71,22 @@ class TestWhenValidInput:
     ) -> None:
         # Arrange
         calculation_input_path = f"{str(tmp_path)}/{calculation_input_folder}"
-        table_location = f"{calculation_input_path}/charge_masterdata_periods"
-        row = _create_charge_master_period_row()
-        df = spark.createDataFrame(data=[row], schema=charge_master_data_periods_schema)
+        table_location = f"{calculation_input_path}/charge_price_points"
+        row = _create_change_price_point_row()
+        df = spark.createDataFrame(data=[row], schema=charge_price_points_schema)
         write_dataframe_to_table(
             spark,
             df,
             "test_database",
-            "charge_master_data_periods",
+            "charge_price_points",
             table_location,
-            charge_master_data_periods_schema,
+            charge_price_points_schema,
         )
         expected = df
         reader = TableReader(spark, calculation_input_path)
 
         # Act
-        actual = reader.read_charge_master_data_periods()
+        actual = reader.read_charge_price_points()
 
         # Assert
         assert_dataframes_equal(actual, expected)
