+++ source/databricks/calculation_engine/tests/calculation/input/table_reader/test_read_metering_point_periods.py
@@ -18,9 +18,9 @@ import pytest
 from pyspark.sql import SparkSession
 import pyspark.sql.functions as f
 
-from package.calculation_input.table_reader import TableReader
-from package.calculation_input.schemas import metering_point_period_schema
-import tests.calculation_input.table_reader.input_metering_point_periods_factory as factory
+from package.calculation.input.table_reader import TableReader
+from package.calculation.input import metering_point_period_schema
+import calculation.input.table_reader.input_metering_point_periods_factory as factory
 from tests.helpers.delta_table_utils import write_dataframe_to_table
 from tests.helpers.data_frame_utils import assert_dataframes_equal
