+++ source/databricks/calculation_engine/tests/calculation/input/test_input_table_schemas.py
@@ -16,7 +16,7 @@
 from pyspark.sql import SparkSession
 from pyspark.sql.functions import StructType
 
-from package.calculation_input.schemas import (
+from package.calculation.input import (
     time_series_point_schema,
     metering_point_period_schema,
     charge_link_periods_schema,
