+++ source/databricks/calculation_engine/tests/calculation/preparation/transformations/charge_types/test_tariffs.py
@@ -15,7 +15,6 @@ from datetime import datetime
 from decimal import Decimal
 
 import pytest
-from pyspark.sql.functions import lit
 
 import package.codelists as e
 
@@ -25,9 +24,8 @@ from package.calculation.preparation.transformations import (
     get_tariff_charges,
 )
 from package.calculation.wholesale.schemas.tariffs_schema import tariff_schema
-from package.calculation_input.schemas import (
+from package.calculation.input import (
     time_series_point_schema,
-    charge_price_points_schema,
 )
 
 from package.constants import Colname
