+++ source/databricks/calculation_engine/tests/calculator_job/conftest.py
@@ -22,7 +22,7 @@ import package.calculation as calculation
 from package.calculation.calculator_args import CalculatorArgs
 from package.calculation.preparation import PreparedDataReader
 
-from package.calculation_input import TableReader
+from package.calculation.input import TableReader
 from package.codelists.calculation_type import (
     CalculationType,
 )
