+++ source/databricks/calculation_engine/tests/datamigration/test_grid_loss_metering_points_migrations.py
@@ -13,12 +13,12 @@
 # limitations under the License.
 
 from pyspark.sql import SparkSession, DataFrame
-from pyspark.sql.functions import array, col, lit
+from pyspark.sql.functions import col
 import pytest
 
 from helpers.data_frame_utils import set_column
 
-from package.calculation_input.schemas import grid_loss_metering_points_schema
+from package.calculation.input import grid_loss_metering_points_schema
 from package.infrastructure.paths import (
     GRID_LOSS_METERING_POINTS_TABLE_NAME,
     INPUT_DATABASE_NAME,
