+++ source/databricks/calculation_engine/tests/infrastructure/test_get_calculator_args.py
@@ -56,7 +56,7 @@ def job_environment_variables() -> dict:
     return {
         EnvironmentVariable.TIME_ZONE.name: "Europe/Copenhagen",
         EnvironmentVariable.DATA_STORAGE_ACCOUNT_NAME.name: "some_storage_account_name",
-        EnvironmentVariable.CALCULATION_INPUT_FOLDER_NAME.name: "calculation_input",
+        EnvironmentVariable.CALCULATION_INPUT_FOLDER_NAME.name: "input",
         EnvironmentVariable.TENANT_ID.name: "550e8400-e29b-41d4-a716-446655440000",
         EnvironmentVariable.SPN_APP_ID.name: "some_spn_app_id",
         EnvironmentVariable.SPN_APP_SECRET.name: "some_spn_app_secret",
@@ -116,7 +116,7 @@ class TestWhenInvokedWithValidParameters:
         # Assert - Infrastructure settings
         assert (
             actual_settings.calculation_input_path
-            == "abfss://wholesale@some_storage_account_name.dfs.core.windows.net/calculation_input/"
+            == "abfss://wholesale@some_storage_account_name.dfs.core.windows.net/input/"
         )
         assert (
             actual_settings.wholesale_container_path
@@ -124,7 +124,7 @@ class TestWhenInvokedWithValidParameters:
         )
         assert (
             actual_settings.calculation_input_path
-            == "abfss://wholesale@some_storage_account_name.dfs.core.windows.net/calculation_input/"
+            == "abfss://wholesale@some_storage_account_name.dfs.core.windows.net/input/"
         )
 
     def test_parses_optional_time_series_points_table_name(
